feat: rename getstripe file
